refactor: 讓 Profile Save 流程收斂成一個深模組 (X-Tracker #352)
- 建立 src/lib/profile/saveProfile.ts 純 TS 深模組，封裝 7 步驟儲存編排邏輯，避免直接依賴 React 或 Next.js hooks
- 在 .eslintrc.json 中配置
o-restricted-imports 確保 saveProfile.ts 維持純粹性
- 重構 useProfileSubmit.ts 為極簡 Hook，僅負責 UI 的錯誤捲動與 isSaving 狀態控制
- 拆分測試為 saveProfile.test.ts（涵蓋 7 步驟儲存與異步協調）與 useProfileSubmit.test.ts（涵蓋 UI 防禦狀態）
- 確保所有測試與類型檢查皆 100% 通過

Ref: https://github.com/Xchange-Taiwan/X-Talent-Tracker/issues/352
